requirements.txt
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coverage==5.2
cryptography==2.9.2
dropbox==10.2.0
Flask==1.1.2
Flask-API==2.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mock==4.0.2
more-itertools==8.4.0
numpy==1.19.0
packaging==20.4
pandas==1.0.5
Pillow==7.2.0
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
xlrd==1.2.0
zipp==3.1.0
